Flatten calculate_offsets directory tree
In a move to remove references to 'rfc003' flatten the directory tree
under `src/bin/calculate_offsets/` to remove the `rfc003` directories.
