+++ src/bin/calculate_offsets/bitcoin/templates/bitcoin/config.json
