+++ src/bin/calculate_offsets/bitcoin/templates/bitcoin/contract.script
