+++ src/bin/calculate_offsets/ethereum/templates/erc20/config.json
