+++ src/bin/calculate_offsets/ethereum/templates/erc20/contract.asm
