+++ src/bin/calculate_offsets/ethereum/templates/erc20/deploy_header.asm
