+++ src/bin/calculate_offsets/ethereum/templates/ether/config.json
