+++ src/bin/calculate_offsets/ethereum/templates/ether/contract.asm
