+++ src/bin/print_offsets.rs
@@ -16,16 +16,14 @@
 mod calculate_offsets;
 
 use self::calculate_offsets::{
-    bitcoin::rfc003::BitcoinScript, ethereum::rfc003::contract::EthereumContract,
-    offset::to_markdown, Contract,
+    bitcoin::BitcoinScript, ethereum::contract::EthereumContract, offset::to_markdown, Contract,
 };
 use crate::calculate_offsets::placeholder_offsets;
 use std::ffi::OsStr;
 
-const ETHER_TEMPLATE_FOLDER: &str = "./src/bin/calculate_offsets/ethereum/rfc003/templates/ether/";
-const ERC20_TEMPLATE_FOLDER: &str = "./src/bin/calculate_offsets/ethereum/rfc003/templates/erc20/";
-const BITCOIN_TEMPLATE_FOLDER: &str =
-    "./src/bin/calculate_offsets/bitcoin/rfc003/templates/bitcoin/";
+const ETHER_TEMPLATE_FOLDER: &str = "./src/bin/calculate_offsets/ethereum/templates/ether/";
+const ERC20_TEMPLATE_FOLDER: &str = "./src/bin/calculate_offsets/ethereum/templates/erc20/";
+const BITCOIN_TEMPLATE_FOLDER: &str = "./src/bin/calculate_offsets/bitcoin/templates/bitcoin/";
 
 #[allow(clippy::print_stdout)]
 fn main() -> Result<(), Error> {
@@ -62,18 +60,18 @@ fn generate_markdown<C: Contract, S: AsRef<OsStr>>(template_folder: S) -> Result
 
 #[derive(Debug)]
 enum Error {
-    BitcoinScript(self::calculate_offsets::bitcoin::rfc003::Error),
-    EthereumContract(self::calculate_offsets::ethereum::rfc003::Error),
+    BitcoinScript(self::calculate_offsets::bitcoin::Error),
+    EthereumContract(self::calculate_offsets::ethereum::Error),
 }
 
-impl From<self::calculate_offsets::bitcoin::rfc003::Error> for Error {
-    fn from(err: self::calculate_offsets::bitcoin::rfc003::Error) -> Self {
+impl From<self::calculate_offsets::bitcoin::Error> for Error {
+    fn from(err: self::calculate_offsets::bitcoin::Error) -> Self {
         Error::BitcoinScript(err)
     }
 }
 
-impl From<self::calculate_offsets::ethereum::rfc003::Error> for Error {
-    fn from(err: self::calculate_offsets::ethereum::rfc003::Error) -> Self {
+impl From<self::calculate_offsets::ethereum::Error> for Error {
+    fn from(err: self::calculate_offsets::ethereum::Error) -> Self {
         Error::EthereumContract(err)
     }
 }
